+++ tests/dsl/elements/test_null.py
@@ -5,7 +5,8 @@ import pytest
 from statham.dsl.constants import NotPassed
 from statham.dsl.elements import Null
 from tests.dsl.elements.helpers import assert_validation
-from tests.helpers import Args, no_raise
+from statham.dsl.helpers import Args
+from tests.helpers import no_raise
 
 
 class TestNullInstantiation: